Refactoring & remove unused code
